feat(linkify): modernize linkify code (#1470)
